pyproject.toml
[project]
name = "hiho-cli-audio"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.24.0",
    "httpx>=0.28.1",
    "numpy>=2.3.1",
    "pydantic>=2.11.7",
    "pynput>=1.8.1",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "sounddevice>=0.5.2",
    "typer>=0.16.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.1",
]

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "W", "UP", "D", "D213", "PT"]
ignore = [
    "E501",  # line-too-long
    "D400",  # missing-trailing-period。日本語の「。」に対応していないため。
    "D403",  # first-word-uncapitalized。日本語とは無関係であるため。日英混合時の挙動に一貫性が無いため。
]
unfixable = [
    "F401",  # unused-import
    "F841",  # unused-variable
]

[tool.ruff.lint.per-file-ignores]
"test/**/test_*.py" = ["D103"]  # undocumented-public-function。pytest の対象 test_*() が D103 の対象になり、関数名から内容が自明でも docstring を強制されるため。

[tool.ruff.lint.isort]
known-first-party = ["hiho-cli-audio"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.format]
docstring-code-format = true
